+++ packages/react-ui/components/ComboBox/__stories__/Combobox.stories.tsx
@@ -1,3 +1,5 @@
+// TODO: Rewrite stories and enable rule (in process of functional refactoring).
+/* eslint-disable react/no-unstable-nested-components */
 import React from 'react';
 import { action } from '@storybook/addon-actions';
 import BabyIcon from '@skbkontur/react-icons/Baby';
@@ -36,6 +38,8 @@ SimpleComboboxStory.parameters = {
         in: ['ie11', 'ie118px', 'ie11Flat8px', 'ie11Dark'],
         tests: ['hovered', 'selected_2', 'select_1'],
       },
+      // TODO @Khlutkova fix after update browsers
+      { in: ['chrome8px', 'chromeFlat8px', 'chrome', 'chromeDark'], tests: ['hovered'] },
     ],
     tests: {
       async plain() {
@@ -89,7 +93,7 @@ SimpleComboboxStory.parameters = {
           .perform();
         await this.expect(await this.takeScreenshot()).to.matchImage('selected');
       },
-      async ['search result']() {
+      async 'search result'() {
         await this.browser
           .actions({
             bridge: true,
@@ -127,7 +131,7 @@ SimpleComboboxStory.parameters = {
           .perform();
         await this.expect(await this.takeScreenshot()).to.matchImage('selcted');
       },
-      async ['opened again']() {
+      async 'opened again'() {
         await this.browser
           .actions({
             bridge: true,
@@ -155,7 +159,7 @@ SimpleComboboxStory.parameters = {
           .perform();
         await this.expect(await this.takeScreenshot()).to.matchImage('opened again');
       },
-      async ['search result_0']() {
+      async 'search result_0'() {
         await this.browser
           .actions({
             bridge: true,
@@ -280,7 +284,11 @@ OpenToTop.storyName = 'open to top';
 
 OpenToTop.parameters = {
   creevey: {
-    skip: [{ in: ['ie11', 'ie118px', 'ie11Flat8px', 'ie11Dark'], tests: 'hovered' }],
+    skip: [
+      { in: ['ie11', 'ie118px', 'ie11Flat8px', 'ie11Dark'], tests: 'hovered' },
+      // TODO @Khlutkova fix after update browsers
+      { in: ['chrome8px', 'chromeFlat8px', 'chrome', 'chromeDark'], tests: ['hovered'] },
+    ],
     tests: {
       async plain() {
         const element = await this.browser.findElement({ css: '[data-tid="container"]' });
@@ -482,25 +490,21 @@ export const WithBorderless = () => (
 WithBorderless.storyName = 'with borderless';
 WithBorderless.parameters = { creevey: { skip: [true] } };
 
-export const WithCenterAlign = () => (
-  <SimpleCombobox align={'center'} placeholder={'placeholder'} noInitialValue={true} />
-);
+export const WithCenterAlign = () => <SimpleCombobox align={'center'} placeholder={'placeholder'} noInitialValue />;
 WithCenterAlign.storyName = 'with center align';
 WithCenterAlign.parameters = { creevey: { skip: [true] } };
 
 export const NotRenderNotFound = () => (
-  <SimpleCombobox placeholder={'placeholder'} noInitialValue={true} renderNotFound={() => null} />
+  <SimpleCombobox placeholder={'placeholder'} noInitialValue renderNotFound={() => null} />
 );
 NotRenderNotFound.storyName = 'not render NotFound';
 NotRenderNotFound.parameters = { creevey: { skip: [true] } };
 
-export const WithRightAlign = () => (
-  <SimpleCombobox align={'right'} placeholder={'placeholder'} noInitialValue={true} />
-);
+export const WithRightAlign = () => <SimpleCombobox align={'right'} placeholder={'placeholder'} noInitialValue />;
 WithRightAlign.storyName = 'with right align';
 WithRightAlign.parameters = { creevey: { skip: [true] } };
 
-export const WithMaxLength = () => <SimpleCombobox maxLength={10} placeholder={'placeholder'} noInitialValue={true} />;
+export const WithMaxLength = () => <SimpleCombobox maxLength={10} placeholder={'placeholder'} noInitialValue />;
 WithMaxLength.storyName = 'with maxLength';
 WithMaxLength.parameters = { creevey: { skip: [true] } };
 
@@ -513,7 +517,7 @@ ToogleError.parameters = {
       async plain() {
         await this.expect(await this.takeScreenshot()).to.matchImage('plain');
       },
-      async ['with error']() {
+      async 'with error'() {
         await this.browser
           .actions({
             bridge: true,
@@ -523,7 +527,7 @@ ToogleError.parameters = {
           .perform();
         await this.expect(await this.takeScreenshot()).to.matchImage('with error');
       },
-      async ['plain again']() {
+      async 'plain again'() {
         await this.browser
           .actions({
             bridge: true,
@@ -556,10 +560,10 @@ WithExternalValue.storyName = 'with external value';
 WithExternalValue.parameters = {
   creevey: {
     tests: {
-      async ['initial value']() {
+      async 'initial value'() {
         await this.expect(await this.takeScreenshot()).to.matchImage('initial value');
       },
-      async ['reset value']() {
+      async 'reset value'() {
         await this.browser
           .actions({
             bridge: true,
@@ -568,7 +572,7 @@ WithExternalValue.parameters = {
           .perform();
         await this.expect(await this.takeScreenshot()).to.matchImage('reset value');
       },
-      async ['set value']() {
+      async 'set value'() {
         await this.browser
           .actions({
             bridge: true,
@@ -614,7 +618,7 @@ OpenCloseSearchMethods.parameters = { creevey: { skip: [true] } };
 
 export const FocusFlow: Story = () => (
   <div>
-    <SimpleCombobox autoFocus={true} />
+    <SimpleCombobox autoFocus />
     <br />
     <br />
     <SimpleCombobox />
@@ -628,7 +632,7 @@ FocusFlow.parameters = {
       async before() {
         await this.expect(await this.takeScreenshot()).to.matchImage('before');
       },
-      async ['after Enter on Item']() {
+      async 'after Enter on Item'() {
         await this.browser
           .actions({
             bridge: true,
@@ -637,7 +641,7 @@ FocusFlow.parameters = {
           .perform();
         await this.expect(await this.takeScreenshot()).to.matchImage('after Enter on Item');
       },
-      async ['after tab to the next field']() {
+      async 'after tab to the next field'() {
         await this.browser
           .actions({
             bridge: true,
@@ -805,6 +809,8 @@ interface SimpleComboboxState {
   value: Nullable<{ value: number; label: string }>;
 }
 
+type ComboboxItem = { value: number; label: string };
+
 @rootNode
 class SimpleCombobox extends React.Component<SimpleComboboxProps & ComboBoxProps<any>, SimpleComboboxState> {
   public static defaultProps = {
@@ -839,11 +845,11 @@ class SimpleCombobox extends React.Component<SimpleComboboxProps & ComboBoxProps
         { value: 6, label: 'Sixth' },
         { value: 7, label: 'A long long long long long long time ago' },
       ].filter((x) => x.label.toLowerCase().includes(query.toLowerCase()) || x.value.toString(10) === query),
-    ).then<Array<{ value: number; label: string }>>(
-      (result) => new Promise((ok) => setTimeout(ok, this.props.delay || 0, result)),
-    );
+    ).then<ComboboxItem[]>((result) => new Promise((ok) => setTimeout(ok, this.props.delay || 0, result)));
 }
 
+type City = { Id: number; City: string };
+type ComboboxSearchResult = { foundItems: City[]; totalCount: number };
 class ComplexCombobox extends React.Component<Omit<ComboBoxProps<any>, 'getItems'>, {}> {
   public static defaultProps = ComboBox.defaultProps;
   public state = {
@@ -871,7 +877,7 @@ class ComplexCombobox extends React.Component<Omit<ComboBoxProps<any>, 'getItems
 
   private getItems = (query: string) => {
     return getCities(query)
-      .then(({ foundItems, totalCount }: { foundItems: Array<{ Id: number; City: string }>; totalCount: number }) => ({
+      .then(({ foundItems, totalCount }: ComboboxSearchResult) => ({
         foundItems: foundItems.map(this.mapCity),
         totalCount,
       }))
@@ -956,7 +962,7 @@ const items: ValueType[] = [
 ];
 
 function search(query: string) {
-  return Promise.resolve(items.filter((x) => ~x.name.toLowerCase().indexOf(query.toLowerCase())));
+  return Promise.resolve(items.filter((x) => x.name.toLowerCase().indexOf(query.toLowerCase()) !== -1));
 }
 
 let searchCount = 0;
@@ -972,7 +978,7 @@ function searchWithRejections(query: string): Promise<ValueType[]> {
       if (searchCount % 2) {
         throw new Error();
       }
-      return items.filter((x) => ~x.name.indexOf(query.toLowerCase()));
+      return items.filter((x) => x.name.indexOf(query.toLowerCase()) !== -1);
     });
 }
 
@@ -1117,7 +1123,7 @@ WithTooltip.storyName = 'with tooltip';
 WithTooltip.parameters = {
   creevey: {
     tests: {
-      async ['show and hide Tooltip']() {
+      async 'show and hide Tooltip'() {
         const body = await this.browser.findElement({ css: 'body' });
 
         await this.browser