+++ packages/react-ui/components/Modal/Modal.tsx
@@ -2,6 +2,7 @@ import React from 'react';
 import FocusLock from 'react-focus-lock';
 import throttle from 'lodash.throttle';
 
+import { isNonNullable } from '../../lib/utils';
 import { isKeyEscape } from '../../lib/events/keyboard/identifiers';
 import * as LayoutEvents from '../../lib/LayoutEvents';
 import { RenderContainer } from '../../internal/RenderContainer';
@@ -134,7 +135,7 @@ export class Modal extends React.Component<ModalProps, ModalState> {
     }
 
     window.removeEventListener('keydown', this.handleKeyDown);
-    if (this.stackSubscription != null) {
+    if (isNonNullable(this.stackSubscription)) {
       this.stackSubscription.remove();
     }
     ModalStack.remove(this);
@@ -270,7 +271,7 @@ export class Modal extends React.Component<ModalProps, ModalState> {
     this.containerNode = center;
   };
 
-  private handleStackChange = (stack: ReadonlyArray<React.Component>) => {
+  private handleStackChange = (stack: readonly React.Component[]) => {
     this.setState({ stackPosition: stack.indexOf(this), hasBackground: ModalStack.isBlocking(this) });
   };