+++ packages/react-ui/components/Modal/ModalFooter.tsx
@@ -30,7 +30,7 @@ function ModalFooter(props: ModalFooterProps) {
   const modal = useContext(ModalContext);
   const layout = useResponsiveLayout();
 
-  const { sticky = layout.isMobile ? false : true, panel, children } = props;
+  const { sticky = !layout.isMobile, panel, children } = props;
 
   useEffect(() => {
     modal.setHasFooter?.();
@@ -40,7 +40,6 @@ function ModalFooter(props: ModalFooterProps) {
       modal.setHasFooter?.(false);
       modal.setHasPanel?.(false);
     };
-    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [panel]);
 
   const renderContent = (fixed = false) => {