+++ packages/react-ui/components/Paging/Paging.tsx
@@ -18,6 +18,7 @@ import { styles } from './Paging.styles';
 import * as NavigationHelper from './NavigationHelper';
 import { getItems } from './PagingHelper';
 import { PagingLocale, PagingLocaleHelper } from './locale';
+import { PagingDefaultComponent } from './PagingDefaultComponent';
 
 const IGNORE_EVENT_TAGS = ['input', 'textarea'];
 
@@ -54,6 +55,15 @@ export interface PagingProps extends CommonProps {
    * на каждом из них. Такие случаи лучше обрабатывать отдельно.
    */
   useGlobalListener: boolean;
+  /**
+   * Определяет, нужно ли показывать `Paging` когда страница всего одна.
+   *
+   * Этот проп будет удалён в 5-ой версии библиотеки,
+   * так как поведение со скрытием `Paging`'а станет поведением по умолчанию.
+   *
+   * @default false
+   */
+  shouldBeVisibleWithLessThanTwoPages: boolean;
 }
 
 export interface PagingState {
@@ -70,13 +80,10 @@ export class Paging extends React.PureComponent<PagingProps, PagingState> {
   public static __KONTUR_REACT_UI__ = 'Paging';
 
   public static defaultProps = {
-    component: ({ className, onClick, children }: any) => (
-      <span className={className} onClick={onClick}>
-        {children}
-      </span>
-    ),
+    component: PagingDefaultComponent,
+    shouldBeVisibleWithLessThanTwoPages: true,
     useGlobalListener: false,
-    ['data-tid']: 'Paging__root',
+    'data-tid': 'Paging__root',
   };
 
   public static propTypes = {};
@@ -125,6 +132,10 @@ export class Paging extends React.PureComponent<PagingProps, PagingState> {
   }
 
   public render() {
+    if (this.props.pagesCount < 2 && !this.props.shouldBeVisibleWithLessThanTwoPages) {
+      return null;
+    }
+
     return (
       <ThemeContext.Consumer>
         {(theme) => {
@@ -413,7 +424,7 @@ export class Paging extends React.PureComponent<PagingProps, PagingState> {
   };
 
   private goToPage = (pageNumber: number) => {
-    if (1 <= pageNumber && pageNumber !== this.props.activePage && pageNumber <= this.props.pagesCount) {
+    if (pageNumber >= 1 && pageNumber !== this.props.activePage && pageNumber <= this.props.pagesCount) {
       this.props.onPageChange(pageNumber);
     }
   };