+++ packages/react-ui/components/Tabs/__stories__/Tabs.stories.tsx
@@ -1,3 +1,5 @@
+// TODO: Rewrite stories and enable rule (in process of functional refactoring).
+/* eslint-disable react/no-unstable-nested-components */
 import React from 'react';
 import { linkTo } from '@storybook/addon-links';
 
@@ -386,6 +388,7 @@ const tabsTests: CreeveyTests = {
       })
       .sendKeys(this.keys.TAB)
       .perform();
+    await delay(1000);
     await this.expect(await this.takeScreenshot()).to.matchImage('tabPress');
   },
   async enterPress() {
@@ -409,6 +412,7 @@ const tabsTests: CreeveyTests = {
       })
       .sendKeys(this.keys.ENTER)
       .perform();
+    await delay(1000);
     await this.expect(await this.takeScreenshot()).to.matchImage('enterPress');
   },
 };
@@ -418,10 +422,14 @@ Simple.storyName = 'simple';
 
 Simple.parameters = {
   creevey: {
-    skip: [{ in: ['ie11', 'ie118px', 'ie11Dark'], tests: 'hovered' }],
+    skip: [
+      { in: ['ie11', 'ie118px', 'ie11Dark'], tests: 'hovered' },
+      // TODO @Khlutkova fix after update browsers
+      { in: ['chrome8px', 'chromeFlat8px', 'chrome', 'chromeDark'], tests: ['hovered', 'focused', 'tabPress'] },
+    ],
     tests: {
       ...tabsTests,
-      async ['move focus forward']() {
+      async 'move focus forward'() {
         await this.browser
           .actions({
             bridge: true,
@@ -438,7 +446,7 @@ Simple.parameters = {
           .perform();
         await this.expect(await this.takeScreenshot()).to.matchImage('move focus forward');
       },
-      async ['move focus backward']() {
+      async 'move focus backward'() {
         await this.browser
           .actions({
             bridge: true,
@@ -461,7 +469,7 @@ Simple.parameters = {
           .perform();
         await this.expect(await this.takeScreenshot()).to.matchImage('move focus backward');
       },
-      async ['reset focus after click']() {
+      async 'reset focus after click'() {
         await this.browser
           .actions({
             bridge: true,
@@ -511,7 +519,17 @@ HrefsSecond.parameters = { creevey: { skip: [true] } };
 export const Vertical: Story = () => <UncTabs vertical />;
 Vertical.storyName = 'vertical';
 Vertical.parameters = {
-  creevey: { skip: [{ in: ['ie11', 'ie118px', 'ie11Dark'], tests: 'hovered' }], tests: tabsTests },
+  creevey: {
+    skip: [
+      { in: ['ie11', 'ie118px', 'ie11Dark'], tests: 'hovered' },
+      // TODO @Khlutkova fix after update browsers
+      {
+        in: ['chrome8px', 'chromeFlat8px', 'chrome', 'chromeDark'],
+        tests: ['hovered', 'focused', 'tabPress', 'enterPress'],
+      },
+    ],
+    tests: tabsTests,
+  },
 };
 
 export const WithComponent = () => <TabsWithMyLink />;
@@ -525,7 +543,14 @@ WithUnexpectedTabSizeChange.parameters = { creevey: { skip: [true] } };
 export const WithDisabledTab: Story = () => <DisabledTab />;
 WithDisabledTab.storyName = 'with disabled tab';
 WithDisabledTab.parameters = {
-  creevey: { skip: [{ in: ['ie11', 'ie118px', 'ie11Dark'], tests: 'hovered' }], tests: tabsTests },
+  creevey: {
+    skip: [
+      { in: ['ie11', 'ie118px', 'ie11Dark'], tests: 'hovered' },
+      // TODO @Khlutkova fix after update browsers
+      { in: ['chrome8px', 'chromeFlat8px', 'chrome', 'chromeDark'], tests: ['hovered', 'focused', 'tabPress'] },
+    ],
+    tests: tabsTests,
+  },
 };
 
 export const TabsInModalStory = () => <TabsInModal />;