+++ packages/react-ui/components/Toggle/Toggle.styles.ts
@@ -30,8 +30,7 @@ export const styles = memoizeStyle({
         width: ${handleActiveWidth};
       }
       &:active:not(.${globalClasses.disabled}) input:checked ~ .${globalClasses.handle} {
-        transform: translateX(${t.toggleWidth}) translateX(-${handleWidthWithBorders})
-          translateX(-${t.toggleHandleActiveWidthIncrement});
+        transform: translateX(${t.toggleWidth}) translateX(-${handleWidthWithBorders}) translateX(-4px);
       }
 
       &::before {
@@ -47,18 +46,17 @@ export const styles = memoizeStyle({
   },
 
   handle(t: Theme) {
-    const handleSize = `calc(${t.toggleHeight} - 2 * ${t.toggleBorderWidth})`;
     return css`
-      background: ${t.toggleBg};
+      background: ${t.toggleHandleBg};
       border-radius: ${t.toggleHandleBorderRadius};
       bottom: ${t.toggleBorderWidth};
       box-shadow: 0 ${t.toggleBorderWidth} 0 0 rgba(0, 0, 0, 0.15), 0 0 0 ${t.toggleBorderWidth} rgba(0, 0, 0, 0.15);
-      height: ${handleSize};
-      left: ${t.toggleBorderWidth};
+      height: ${t.toggleHandleSize};
+      left: ${t.toggleHandleLeft};
       position: absolute;
-      top: ${t.toggleBorderWidth};
+      top: ${t.toggleHandleTop};
       transition: 0.2s ease-in;
-      width: ${handleSize};
+      width: ${t.toggleHandleSize};
     `;
   },
 
@@ -103,6 +101,10 @@ export const styles = memoizeStyle({
       }
       &:checked ~ .${globalClasses.handle} {
         transform: translateX(${t.toggleWidth}) translateX(-${handleWidthWithBorders});
+        background: ${t.toggleCheckedBg};
+        &:hover {
+          background: ${t.toggleCheckedBgHover};
+        }
       }
     `;
   },