+++ packages/react-ui/internal/FileUploaderControl/withFileUploaderControlProvider.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable react/display-name */
 import React, { ComponentType } from 'react';
 
 import { FileUploaderControlProvider, FileUploaderControlProviderProps } from './FileUploaderControlProvider';
@@ -6,8 +7,8 @@ export const withFileUploaderControlProvider = <TProps extends object, TRef exte
   Component: ComponentType<
     Pick<TProps & FileUploaderControlProviderProps, Exclude<keyof TProps, keyof FileUploaderControlProviderProps>>
   >,
-) =>
-  React.memo(
+) => {
+  return React.memo(
     React.forwardRef<TRef, TProps & FileUploaderControlProviderProps>(
       (props: TProps & FileUploaderControlProviderProps, ref) => {
         const { onRemove, onValueChange, onAttach, ...rest } = props;
@@ -19,3 +20,4 @@ export const withFileUploaderControlProvider = <TProps extends object, TRef exte
       },
     ),
   );
+};