+++ packages/react-ui/lib/theming/ThemeFactory.ts
@@ -1,6 +1,8 @@
 import { DefaultThemeInternal } from '../../internal/themes/DefaultTheme';
+import { isNonNullable } from '../utils';
 
 import { Theme, ThemeIn } from './Theme';
+import { findPropertyDescriptor } from './ThemeHelpers';
 
 export class ThemeFactory {
   public static create<T extends {}>(theme: ThemeIn & T, baseTheme?: Theme): Readonly<Theme & T> {
@@ -8,23 +10,26 @@ export class ThemeFactory {
     return this.constructTheme(base, theme);
   }
 
-  public static overrideDefaultTheme(theme: ThemeIn) {
-    Object.keys(theme).forEach((variableName) => {
-      const descriptor = Object.getOwnPropertyDescriptor(theme, variableName)!;
+  public static overrideDefaultTheme(theme: Theme) {
+    ThemeFactory.getKeys(DefaultThemeInternal).forEach((variableName) => {
+      const descriptor = findPropertyDescriptor(theme, variableName);
       Object.defineProperty(DefaultThemeInternal, variableName, descriptor);
     });
   }
 
   public static getKeys<T extends Theme>(theme: T) {
     const keys: Array<keyof T> = [];
-    while (theme != null) {
-      (Object.keys(theme) as Array<keyof T>).forEach((key) => {
+    while (isNonNullable(theme)) {
+      (Object.keys(theme) as typeof keys).forEach((key) => {
         if (!keys.includes(key)) {
           keys.push(key);
         }
       });
+      // TODO: Enable `no-param-reassign` rule.
+      // eslint-disable-next-line no-param-reassign
       theme = Object.getPrototypeOf(theme);
     }
+
     return keys.sort();
   }