+++ packages/react-ui-validations/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.8.3](https://github.com/skbkontur/retail-ui/tree/master/packages/react-ui-validations/compare/react-ui-validations@1.8.2...react-ui-validations@1.8.3) (2022-05-04)
+
+**Note:** Version bump only for package react-ui-validations
+
+
+
+
+
+## [1.8.2](https://github.com/skbkontur/retail-ui/tree/master/packages/react-ui-validations/compare/react-ui-validations@1.8.1...react-ui-validations@1.8.2) (2022-04-25)
+
+
+### Bug Fixes
+
+* **validations:** add focus on submit for fields with warning ([#2804](https://github.com/skbkontur/retail-ui/tree/master/packages/react-ui-validations/issues/2804)) ([5b7d468](https://github.com/skbkontur/retail-ui/tree/master/packages/react-ui-validations/commit/5b7d468989f358ad65e9cb484fe3243c5752d333))
+
+
+
+
+
 ## [1.8.1](https://github.com/skbkontur/retail-ui/tree/master/packages/react-ui-validations/compare/react-ui-validations@1.8.0...react-ui-validations@1.8.1) (2022-03-18)